refactored styling of the Homepage
